Add project detail modal
A modal for displaying project details was added to the
project index view to allow users to view details about a
project they either own or are collaborating with.

A project details module was also added to project.js for handling the click event and updating the project detail
modal with payload data from the ajax request.